ts: Swap to new Query implementation.
Swap to new query implementation, removing the old query implementation
completely.

Release note: None
