+++ pkg/storage/ts_maintenance_queue_test.go
@@ -237,10 +237,7 @@ func TestTimeSeriesMaintenanceQueueServer(t *testing.T) {
 	// periods; this simplifies verification.
 	seriesName := "test.metric"
 	sourceName := "source1"
-	// "now" is five minutes in the past to avoid any sort of shenanigans with the
-	// various adjustments we make in the very-recent-past to create consistent
-	// graphs.
-	now := tsrv.Clock().PhysicalNow() - int64(5*time.Minute)
+	now := tsrv.Clock().PhysicalNow()
 	nearPast := now - (tsdb.PruneThreshold(ts.Resolution10s) * 2)
 	farPast := now - (tsdb.PruneThreshold(ts.Resolution10s) * 4)
 	sampleDuration := ts.Resolution10s.SampleDuration()
@@ -294,7 +291,7 @@ func TestTimeSeriesMaintenanceQueueServer(t *testing.T) {
 		&memMon,
 		&memMon,
 		ts.QueryMemoryOptions{
-			BudgetBytes:             math.MaxInt64,
+			BudgetBytes:             math.MaxInt64 / 8,
 			EstimatedSources:        1,
 			InterpolationLimitNanos: 0,
 		},
@@ -312,6 +309,7 @@ func TestTimeSeriesMaintenanceQueueServer(t *testing.T) {
 				SampleDurationNanos: ts.Resolution10s.SampleDuration(),
 				StartNanos:          0,
 				EndNanos:            now + ts.Resolution10s.SlabDuration(),
+				NowNanos:            now + (10 * time.Hour).Nanoseconds(),
 			},
 			memContext,
 		)
